WIP: get quit dialog box working (doesn't do anything yet)
