+++ src/Swarm/TUI/View.hs
@@ -57,11 +57,10 @@ import Text.Wrap
 import Brick hiding (Direction)
 import Brick.Focus
 import Brick.Forms
-import Brick.Widgets.Border (hBorder, hBorderWithLabel, joinableBorder, vBorder)
+import Brick.Widgets.Border (borderAttr, hBorder, hBorderWithLabel, joinableBorder, vBorder)
 import Brick.Widgets.Center (center, hCenter)
 import Brick.Widgets.Dialog
 import qualified Brick.Widgets.List as BL
-import qualified Brick.Widgets.Table as BT
 
 import Swarm.Game.Display
 import Swarm.Game.Entity as E
@@ -167,10 +166,6 @@ chooseCursor s locs = case s ^. uiState . uiModal of
   Nothing -> showFirstCursor s locs
   Just _ -> Nothing
 
--- | Width cap for modal and error message windows
-maxModalWindowWidth :: Int
-maxModalWindowWidth = 500
-
 -- | Render the error dialog window with a given error message
 renderErrorDialog :: Text -> Widget Name
 renderErrorDialog err = renderDialog (dialog (Just "Error") Nothing (maxModalWindowWidth `min` requiredWidth)) errContent
@@ -178,59 +173,13 @@ renderErrorDialog err = renderDialog (dialog (Just "Error") Nothing (maxModalWin
   errContent = txtWrapWith indent2 {preserveIndentation = True} err
   requiredWidth = 2 + maximum (textWidth <$> T.lines err)
 
--- | Render a fullscreen widget with some padding
+-- | Render a fullscreen modal widget.
 renderModal :: Modal -> Widget Name
-renderModal modal = renderDialog (dialog (Just modalTitle) modalButtons (maxModalWindowWidth `min` requiredWidth)) modalWidget
- where
-  (modalTitle, modalWidget, modalButtons, requiredWidth) =
-    case modal of
-      HelpModal -> ("Help", helpWidget, Nothing, maxModalWindowWidth)
-      WinModal -> ("", txt "Congratulations!", Nothing, maxModalWindowWidth)
-      QuitModal ->
-        let quitMsg = "Are you sure you want to quit?"
-         in ( ""
-            , padBottom (Pad 1) $ hCenter $ txt quitMsg
-            , Just (0, [("Cancel", False), ("Quit", True)])
-            , T.length quitMsg + 4
-            )
-
-helpWidget :: Widget Name
-helpWidget = (helpKeys <=> fill ' ') <+> (helpCommands <=> fill ' ')
- where
-  helpKeys =
-    vBox
-      [ hCenter $ txt "Global Keybindings"
-      , hCenter $ mkTable glKeyBindings
-      ]
-  mkTable = BT.renderTable . BT.table . map toWidgets
-  toWidgets (k, v) = [txt k, txt v]
-  glKeyBindings =
-    [ ("F1", "Help")
-    , ("Ctrl-q", "quit the game")
-    , ("Tab", "cycle panel focus")
-    , ("Meta-w", "focus on the world map")
-    , ("Meta-e", "focus on the robot inventory")
-    , ("Meta-r", "focus on the REPL")
-    , ("Meta-t", "focus on the info panel")
-    ]
-  helpCommands =
-    vBox
-      [ hCenter $ txt "Commands"
-      , hCenter $ mkTable baseCommands
-      ]
-  baseCommands =
-    [ ("build <name> {<commands>}", "Create a robot")
-    , ("make <name>", "Craft an item")
-    , ("move", "Move one step in the current direction")
-    , ("turn <dir>", "Change the current direction")
-    , ("grab", "Grab whatver is available")
-    , ("give <robot> <item>", "Give an item to another robot")
-    , ("has <item>", "Check for an item in the inventory")
-    ]
+renderModal (Modal _ d w) = renderDialog d w
 
 -- | Draw the error dialog window, if it should be displayed right now.
 drawDialog :: UIState -> Widget Name
-drawDialog s = case s ^. uiModal of
+drawDialog s = overrideAttr borderAttr highlightAttr $ case s ^. uiModal of
   Just m -> renderModal m
   Nothing -> maybe emptyWidget renderErrorDialog (s ^. uiError)
